// src/ui/web/src/types/runner.ts
export interface Runner {
  id: number;
  name: string;
  clubId: number;
  club: string;
  classId: number;
  class: string;
  cardNo: number;
  bib: string;
  startTime: number;
  finishTime: number;
  runningTime: number;
  status: number;
  birthYear: number;
  nationality: string;
  sex: number;
}